posting-function: <HeaderA>というコンポーネントを新たに作り、リファクタリングを行いました。
