+++ package.json
@@ -54,6 +54,8 @@
   },
   "devDependencies": {
     "@testing-library/dom": "^8.0.0",
-    "@testing-library/user-event": "^13.1.9"
+    "@testing-library/user-event": "^13.1.9",
+    "jest-dom": "^4.0.0",
+    "react-test-render": "^1.1.2"
   }
 }
