--- src/components/CreatePost/PostPreview.tsx
@@ -1,211 +1,0 @@
-import React, { useState } from "react";
-import { useSelector } from "react-redux";
-import { selectUser } from "../../features/userSlice";
-import { storage, db } from "../../firebase";
-import firebase from "firebase/app";
-import DefaultButton from "../Parts/InputFileButton";
-import SecondaryButton from "../Parts/SecondaryButton";
-import styled from "styled-components";
-import mediaQuery from "styled-media-query";
-import {
-  Paper,
-  makeStyles,
-  createStyles,
-  Theme,
-  Slide,
-} from "@material-ui/core";
-import { ArrowDownward } from "@material-ui/icons";
-
-const mediaMobile = mediaQuery.lessThan("medium");
-
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    paper: {
-      width: "100%",
-      height: "100%",
-    },
-    paperForPreview: {
-      width: "100%",
-      height: "100%",
-      position: "absolute",
-      top: "0",
-      left: "0",
-    },
-  })
-);
-
-const Wrapper = styled.div`
-  position: relative;
-  width: 30vw;
-  ${mediaMobile`
-  width: 100vw;`}
-  height: calc(100vh - 40px);
-  ${mediaMobile`
-  height: 100vh`};
-  margin: 20px auto;
-  ${mediaMobile`
-  margin: 0;`}
-  padding: 0;
-`;
-
-const Header = styled.div`
-  display: flex;
-  width: 100%;
-  height: 52px;
-  margin: 0 auto;
-  background-color: hsl(0, 0%, 100%);
-  border-bottom: 1px solid hsla(26, 100%, 12%, 0.2);
-  box-sizing: border-box;
-`;
-
-const Title = styled.h2`
-  width: 100%;
-  height: 52px;
-  margin: 0 auto;
-  font-size: 18px;
-  line-height: 52px;
-  text-align: center;
-  color: hsl(0, 0%, 10%);
-  font-weight: bold;
-  letter-spacing: 2px;
-`;
-
-const Main = styled.main`
-  display: flex;
-  flex-flow: column;
-  width: 100%;
-  margin: 0 auto;
-  background-color: hsl(0, 0%, 100%);
-`;
-
-const UserInfo = styled.div`
-  display: flex;
-  width: 100%;
-  height: 60px;
-  margin: 0 auto;
-  background-color: hsl(0, 0%, 100%);
-  border-bottom: 1px solid hsla(26, 100%, 12%, 0.2);
-  box-sizing: border-box;
-`;
-
-const UserIcon = styled.div`
-  width: 40px;
-  height: 40px;
-  margin: 10px;
-  border-radius: 100%;
-  background-color: hsl(0, 0%, 90%);
-`;
-
-const UserImage = styled.img`
-  width: 100%;
-  height: 100%;
-  margin: 0;
-  padding: 0;
-  border-radius: 100%;
-`;
-
-const UserName = styled.p`
-  width: 150px;
-  height: 60px;
-  margin: 0;
-  line-height: 60px;
-`;
-
-const ImageWrap = styled.div`
-  width: 100%;
-  margin: 0 auto;
-  position: relative;
-  :before {
-    content: "";
-    display: block;
-    padding-top: 75%;
-  }
-`;
-
-const Image = styled.img`
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  top: 0;
-  left: 0;
-  object-fit: cover;
-`;
-
-const NoImage = styled(Image)`
-  opacity: 0.3;
-`;
-
-const Notes = styled.div`
-  position: absolute;
-  left: 50%;
-  bottom: 20px;
-  ${mediaMobile`
-  bottom: 0px;`}
-  width: 200px;
-  height: 60px;
-  transform: translateX(-50%);
-  text-align: center;
-`;
-
-const ButtonArea = styled.div`
-  display: flex;
-  width: 100%;
-  height: 40px;
-  margin-top: 20px;
-  justify-content: space-around;
-`;
-
-const Textarea = styled.textarea`
-  display: block;
-  width: 80%;
-  height: 100px;
-  margin: 20px auto 0;
-  padding: 10px;
-  border-radius: 10px;
-  border: none;
-  font-size: 1rem;
-  resize: none;
-  background-color: hsl(0, 0%, 95%);
-  font-family: "Yu Gothic";
-`;
-
-const CommentArea = styled.p`
-  width: 80%;
-  height: 100px;
-  margin: 20px auto 0;
-  padding: 10px;
-  border-radius: 10px;
-  border: none;
-  font-size: 1rem;
-`;
-
-type Props = {
-  in: boolean;
-};
-
-const PostPreview = (props: Props) => {
-  const user = useSelector(selectUser);
-  const noImage = `${process.env.PUBLIC_URL}/noPhoto.png`;
-  const [imageUrl, setImageUrl] = useState<string>(noImage);
-  const [caption, setCaption] = useState<string>("");
-  const [preview, setPrevie] = useState<boolean>(false);
-
-  const classes = useStyles();
-
-  return (
-    <>
-      <Slide direction="left" in={props.in} mountOnEnter unmountOnExit>
-        <Paper elevation={2} className={classes.paperForPreview} data-testid="paperForPreview">
-          <Main>
-            <Header data-testid="previewHeader">
-              <Title data-testid="previewTitle">
-                この内容で登録しますか？
-              </Title>
-            </Header>
-          </Main>
-        </Paper>
-      </Slide>
-    </>
-  );
-};
-export default PostPreview;
